+++ src/ciphers.py
@@ -1,26 +1,20 @@
-#!/usr/bin/env python3
-# This tool decodes/encodes various ciphers 
-
-import argparse
-
-
 # ---- Ciphers ----
 
-# OG Caesar - shift of 3 
+# OG Caesar - shift of 3
 def caesar(strng, encode=True):
     az = 'abcdefghijklmnopqrstuvwxyz'
     new = ''
     if encode:
         for i in strng:
             if i.isalpha():
-                 new += az[(az.index(i) + 3) % 26]
+                new += az[(az.index(i) + 3) % 26]
             else:
                 new += i
     else:
         for i in strng:
             if i.isalpha():
                 if az.index(i) < 3:
-                     new += az[26 - (abs(az.index(i) - 3) % 26)]
+                    new += az[26 - (abs(az.index(i) - 3) % 26)]
                 else:
                     new += az[az.index(i) - 3]
             else:
@@ -137,7 +131,7 @@ def playfair(strng, key, encode=True):
     # Append last row to diagram
     diagram.append(row)
 
-    # -----Format Strng----- 
+    # -----Format Strng-----
 
     # Convert strng to upper + no space
     strng = strng.upper()
@@ -208,73 +202,12 @@ def playfair(strng, key, encode=True):
                     encoded.append(diagram[coord_b[4]][coord_b[1]])
                 else:
                     encoded.append(diagram[coord_b[0] - 1][coord_b[1]])
-        # Rule 3: Make square and grab opposite corner on same row        
+        # Rule 3: Make square and grab opposite corner on same row
         else:
             encoded.append(diagram[coord_a[0]][coord_b[1]])
             encoded.append(diagram[coord_b[0]][coord_a[1]])
 
-    # -----Return-----                                                    
+    # -----Return-----
     return ''.join(encoded)
 
-
-# ---- Function Map ----
-
-FUNCTION_MAP = {'3':caesar,
-                '47':rot47,
-                '13':rot13,
-                '5':vigenere,
-                '11':playfair,}
-
-
-
-# ---- Function Names ----
-
-def func_names(a):
-    FUNCTION_NAMES = {'3':'Caesar',
-                      '13':'Rot-13',
-                      '47':'Rot-47',
-                      '5':'Vigenere',
-                      '11':'Playfair'}
-    return FUNCTION_NAMES[a]
-
-# ---- Arguments ----
-
-parser = argparse.ArgumentParser(description='Encrypt a string.')
-parser.add_argument('-c','--code', required=True, choices=FUNCTION_MAP.keys(), metavar='', help='Encrypt message using the given cipher')
-parser.add_argument('-s','--string', required=True, metavar='', help='String to be Encrypted/Decrypted')
-parser.add_argument('-k','--key', metavar='', help='Key/Password for the cipher')
-en_de = parser.add_mutually_exclusive_group()
-en_de.add_argument('-D', '--decode', action='store_true', help='decode the string using current cipher')
-en_de.add_argument('-E', '--encode', action='store_true', help='encode the string using current cipher')
-
-volume = parser.add_mutually_exclusive_group()
-volume.add_argument('-q','--quiet', action='store_true', help='print quiet')
-volume.add_argument('-v','--verbose', action='store_true', help='print verbose')
-
-args = parser.parse_args()
-
-# ---- Handling ----
-
-if __name__ == '__main__':
-    func = FUNCTION_MAP[args.code]
-
-    if args.encode:
-        if args.key:
-            final = func(args.string, args.key, True)
-        else: final = func(args.string, True)
-    elif args.decode:
-        if args.key:
-            final = func(args.string, args.key, False)
-        else: final = func(args.string, False)
-    else:
-        if args.key:
-            final = func(args.string, args.key)
-        else: final = func(args.string)
-
-    if args.quiet:
-        print(final)
-    elif args.verbose:
-        print(f'Using the {func_names(args.code)} Cipher to encrypt \"{args.string}\" results in:\n\n\"{final}\"')
-    else:
-        print(f'The encrypted text is: \"{final}\"')
